+++ .env.example
@@ -1,0 +1,1 @@
+VITE_API_URL=URL_HERE
