+++ package.json
@@ -17,7 +17,6 @@
     "@headlessui/react": "^1.7.17",
     "@tanstack/react-query": "^5.17.12",
     "axios": "^1.6.5",
-    "canvas-txt": "^4.1.1",
     "dayjs": "^1.11.10",
     "eslint-import-resolver-typescript": "^3.6.1",
     "prettier-plugin-organize-imports": "^3.2.4",
