+++ src/utils/gameResult.ts
