--- src/utils/shareResult.ts
@@ -1,17 +1,0 @@
-const shareResult = async (img: string, resultName: string) => {
-  if (!navigator.canShare) {
-    return;
-  }
-  const shareData = {
-    title: `${resultName}!`,
-    text: "Come and check your result at JohnJud!",
-    url: img,
-  };
-  try {
-    await navigator.share(shareData);
-  } catch (error) {
-    console.error("Error sharing", error);
-  }
-};
-
-export { shareResult };
